requirements.txt
aiohttp==3.9.1
asyncio==3.4.3
pandas==2.1.4
pytest==7.4.3
python-dotenv==1.0.0
colorama==0.4.6
requests==2.31.0